Move all js from packs to entrypoints and make sure it loads
