--- .postcssrc.yml
@@ -1,3 +1,0 @@
-plugins:
-  postcss-import: {}
-  postcss-preset-env: {}
