+++ app/javascript/figgy/auth_link_clipboard.js
