+++ app/javascript/figgy/bulk_labeler/labeler_actions_manager.js
@@ -1,4 +1,4 @@
-import lg from "figgy/bulk_labeler/label_generator"
+import lg from '@figgy/bulk_labeler/label_generator'
 export default class LabelerActionsManager {
   constructor(element) {
     this.element = element
