+++ app/javascript/helpers/resource-store-config.js
@@ -1,6 +1,6 @@
 // resource-store-config.spec.js
 import { modules } from 'lux-design-system'
-import { resourceState, resourceMutations, resourceGetters } from "../store/resource"
+import { resourceState, resourceMutations, resourceGetters } from "@store/resource"
 
 export default {
   state: resourceState,
