+++ app/views/layouts/blacklight/base.html.erb
@@ -22,8 +22,7 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.18/dist/css/bootstrap-select.min.css">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.18/dist/js/bootstrap-select.min.js"></script>
 
-    <%# stylesheet_pack_tag 'application' %>
-    <%# javascript_pack_tag "application" %>
+    <%= vite_javascript_tag "application" %>
     <%= content_for(:head) %>
     <%= render 'shared/analytics' if Rails.env.production? %>
   </head>
