+++ app/views/viewer/auth.html.erb
@@ -1,6 +1,5 @@
 <% content_for :head do %>
   <%= stylesheet_link_tag 'application' %>
-  <%# stylesheet_pack_tag 'application' %>
 <% end %>
 <div id="loginContainer">
   <% if @charge_manager.eligible? %>
