+++ app/views/viewer/cdl_checkout.html.erb
@@ -1,6 +1,5 @@
 <% content_for :head do %>
   <%= stylesheet_link_tag 'application' %>
-  <%# stylesheet_pack_tag 'application' %>
 <% end %>
 <div id="cdl-notice">
   <p>This Item may be protected by third-party copyright and/or related intellectual property rights. It is provided by Princeton University Library on a non-commercial basis for your personal fair-use academic and educational research purposes only. For other uses, you must obtain permission from the rights-holder. In addition, you may not make copies of the Item, display it online (e.g., on the internet), or distribute it to anyone else, including friends, colleagues, or classmates. By continuing to use this digital resource you are acknowledging and agree to comply with these terms of use. For additional details, see the following Copyright Notice.</p>
