+++ package.json
@@ -9,6 +9,7 @@
     "@babel/preset-react": "^7.8.3",
     "@google-cloud/pubsub": "^0.31.0",
     "@google-cloud/storage": "^5.8.4",
+    "@vitejs/plugin-vue2": "^2.2.0",
     "ajv": "^6.12.3",
     "apollo-boost": "^0.1.10",
     "axios": "^0.21.2",
@@ -28,6 +29,7 @@
     "react-dom": "^16.12.0",
     "react-redux": "^7.1.3",
     "redux-logger": "^3.0.6",
+    "sass": "^1.62.1",
     "unfetch": "^4.0.1",
     "vue": "^2.5.17",
     "vue-details": "^1.1.0",
